micro-services/rentals/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
pymysql==1.1.1
pika==1.3.2
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
cryptography==42.0.8
firebase-admin==6.5.0
flask-cors==5.0.0

# Testing
pytest==8.2.2
pytest-mock==3.14.0